Split rbm_spin and rbm_multival into header and source files
