+++ CMakeLists.txt
@@ -217,7 +217,8 @@ add_library(netket MODULE Sources/pynetket.cc Sources/Dynamics/py_dynamics.cpp
     Sources/Graph/py_graph.cc Sources/Hilbert/hilbert_index.cc
     Sources/Hilbert/bosons.cc Sources/Hilbert/spins.cc Sources/Hilbert/qubits.cc
     Sources/Hilbert/custom_hilbert.cc Sources/Hilbert/py_hilbert.cc
-    Sources/Machine/abstract_machine.cc)
+    Sources/Machine/abstract_machine.cc Sources/Machine/rbm_spin.cc
+    Sources/Machine/rbm_multival.cc)
 target_link_libraries(netket PUBLIC netket_lib pybind11)
 set_target_properties(netket PROPERTIES PREFIX "${PYTHON_MODULE_PREFIX}"
                                         SUFFIX "${PYTHON_MODULE_EXTENSION}")
