+++ Sources/Machine/rbm_multival.hpp
@@ -18,11 +18,12 @@
 #include <vector>
 #include "Utils/all_utils.hpp"
 #include "Utils/lookup.hpp"
-#include "abstract_machine.hpp"
 #include "rbm_spin.hpp"
 
 #ifndef NETKET_RBM_MULTIVAL_HPP
 #define NETKET_RBM_MULTIVAL_HPP
+
+#include "abstract_machine.hpp"
 
 namespace netket {
 
@@ -69,281 +70,47 @@ class RbmMultival : public AbstractMachine {
 
  public:
   explicit RbmMultival(const AbstractHilbert &hilbert, int nhidden = 0,
-                       int alpha = 0, bool usea = true, bool useb = true)
-      : hilbert_(hilbert),
-        nv_(hilbert.Size()),
-        ls_(hilbert.LocalSize()),
-        usea_(usea),
-        useb_(useb) {
-    nh_ = std::max(nhidden, alpha * nv_);
-    Init();
-  }
+                       int alpha = 0, bool usea = true, bool useb = true);
 
-  void Init() {
-    W_.resize(nv_ * ls_, nh_);
-    a_.resize(nv_ * ls_);
-    b_.resize(nh_);
+  virtual int Npar() const override;
+  virtual int Nvisible() const override;
+  constexpr int Nhidden() const noexcept { return nh_; }
 
-    thetas_.resize(nh_);
-    lnthetas_.resize(nh_);
-    thetasnew_.resize(nh_);
-    lnthetasnew_.resize(nh_);
+  virtual const AbstractHilbert &GetHilbert() const noexcept override;
 
-    npar_ = nv_ * nh_ * ls_;
+  virtual void InitRandomPars(int seed, double sigma) override;
+  virtual void InitLookup(VisibleConstType v, LookupType &lt) override;
+  virtual void UpdateLookup(VisibleConstType v,
+                            const std::vector<int> &tochange,
+                            const std::vector<double> &newconf,
+                            LookupType &lt) override;
 
-    if (usea_) {
-      npar_ += nv_ * ls_;
-    } else {
-      a_.setZero();
-    }
+  virtual VectorType DerLog(VisibleConstType v) override;
+  virtual VectorType DerLog(VisibleConstType v, const LookupType &lt) override;
 
-    if (useb_) {
-      npar_ += nh_;
-    } else {
-      b_.setZero();
-    }
-
-    auto localstates = hilbert_.LocalStates();
-
-    localconfs_.resize(nv_ * ls_);
-    for (int i = 0; i < nv_ * ls_; i += ls_) {
-      for (int j = 0; j < ls_; j++) {
-        localconfs_(i + j) = localstates[j];
-      }
-    }
-
-    mask_.resize(nv_ * ls_, nv_);
-    mask_.setZero();
-
-    for (int i = 0; i < nv_ * ls_; i++) {
-      mask_(i, i / ls_) = 1;
-    }
-
-    for (int i = 0; i < ls_; i++) {
-      confindex_[localstates[i]] = i;
-    }
-
-    vtilde_.resize(nv_ * ls_);
-
-    InfoMessage() << "RBM Multival Initizialized with nvisible = " << nv_
-                  << " and nhidden = " << nh_ << std::endl;
-    InfoMessage() << "Using visible bias = " << usea_ << std::endl;
-    InfoMessage() << "Using hidden bias  = " << useb_ << std::endl;
-    InfoMessage() << "Local size is      = " << ls_ << std::endl;
-  }
-
-  int Nvisible() const override { return nv_; }
-
-  int Nhidden() const { return nh_; }
-
-  int Npar() const override { return npar_; }
-
-  void InitRandomPars(int seed, double sigma) override {
-    VectorType par(npar_);
-
-    netket::RandomGaussian(par, seed, sigma);
-
-    SetParameters(par);
-  }
-
-  void InitLookup(VisibleConstType v, LookupType &lt) override {
-    if (lt.VectorSize() == 0) {
-      lt.AddVector(b_.size());
-    }
-    if (lt.V(0).size() != b_.size()) {
-      lt.V(0).resize(b_.size());
-    }
-    ComputeTheta(v, lt.V(0));
-  }
-
-  void UpdateLookup(VisibleConstType v, const std::vector<int> &tochange,
-                    const std::vector<double> &newconf,
-                    LookupType &lt) override {
-    if (tochange.size() != 0) {
-      for (std::size_t s = 0; s < tochange.size(); s++) {
-        const int sf = tochange[s];
-        const int oldtilde = confindex_[v[sf]];
-        const int newtilde = confindex_[newconf[s]];
-
-        lt.V(0) -= W_.row(ls_ * sf + oldtilde);
-        lt.V(0) += W_.row(ls_ * sf + newtilde);
-      }
-    }
-  }
-
-  VectorType DerLog(VisibleConstType v) override {
-    LookupType ltnew;
-    InitLookup(v, ltnew);
-    return DerLog(v, ltnew);
-  }
-
-  VectorType DerLog(VisibleConstType v, const LookupType &lt) override {
-    VectorType der(npar_);
-    der.setZero();
-
-    ComputeVtilde(v, vtilde_);
-
-    int k = 0;
-
-    if (usea_) {
-      for (; k < nv_ * ls_; k++) {
-        der(k) = vtilde_(k);
-      }
-    }
-
-    RbmSpin::tanh(lt.V(0), lnthetas_);
-
-    if (useb_) {
-      for (int p = 0; p < nh_; p++) {
-        der(k) = lnthetas_(p);
-        k++;
-      }
-    }
-
-    for (int i = 0; i < nv_ * ls_; i++) {
-      for (int j = 0; j < nh_; j++) {
-        der(k) = lnthetas_(j) * vtilde_(i);
-        k++;
-      }
-    }
-    return der;
-  }
-
-  VectorType GetParameters() override {
-    VectorType pars(npar_);
-
-    int k = 0;
-
-    if (usea_) {
-      for (; k < nv_ * ls_; k++) {
-        pars(k) = a_(k);
-      }
-    }
-
-    if (useb_) {
-      for (int p = 0; p < nh_; p++) {
-        pars(k) = b_(p);
-        k++;
-      }
-    }
-
-    for (int i = 0; i < nv_ * ls_; i++) {
-      for (int j = 0; j < nh_; j++) {
-        pars(k) = W_(i, j);
-        k++;
-      }
-    }
-
-    return pars;
-  }
-
-  void SetParameters(VectorConstRefType pars) override {
-    int k = 0;
-
-    if (usea_) {
-      for (; k < nv_ * ls_; k++) {
-        a_(k) = pars(k);
-      }
-    }
-
-    if (useb_) {
-      for (int p = 0; p < nh_; p++) {
-        b_(p) = pars(k);
-        k++;
-      }
-    }
-
-    for (int i = 0; i < nv_ * ls_; i++) {
-      for (int j = 0; j < nh_; j++) {
-        W_(i, j) = pars(k);
-        k++;
-      }
-    }
-  }
+  virtual VectorType GetParameters() override;
+  virtual void SetParameters(VectorConstRefType pars) override;
 
   // Value of the logarithm of the wave-function
-  Complex LogVal(VisibleConstType v) override {
-    ComputeTheta(v, thetas_);
-    RbmSpin::lncosh(thetas_, lnthetas_);
-
-    return (vtilde_.dot(a_) + lnthetas_.sum());
-  }
-
+  virtual Complex LogVal(VisibleConstType v) override;
   // Value of the logarithm of the wave-function
   // using pre-computed look-up tables for efficiency
-  Complex LogVal(VisibleConstType v, const LookupType &lt) override {
-    RbmSpin::lncosh(lt.V(0), lnthetas_);
-
-    ComputeVtilde(v, vtilde_);
-    return (vtilde_.dot(a_) + lnthetas_.sum());
-  }
-
+  virtual Complex LogVal(VisibleConstType v, const LookupType &lt) override;
   // Difference between logarithms of values, when one or more visible variables
   // are being changed
-  VectorType LogValDiff(
+  virtual VectorType LogValDiff(
       VisibleConstType v, const std::vector<std::vector<int>> &tochange,
-      const std::vector<std::vector<double>> &newconf) override {
-    const std::size_t nconn = tochange.size();
-    VectorType logvaldiffs = VectorType::Zero(nconn);
-
-    ComputeTheta(v, thetas_);
-    RbmSpin::lncosh(thetas_, lnthetas_);
-
-    Complex logtsum = lnthetas_.sum();
-
-    for (std::size_t k = 0; k < nconn; k++) {
-      if (tochange[k].size() != 0) {
-        thetasnew_ = thetas_;
-
-        for (std::size_t s = 0; s < tochange[k].size(); s++) {
-          const int sf = tochange[k][s];
-          const int oldtilde = confindex_[v[sf]];
-          const int newtilde = confindex_[newconf[k][s]];
-
-          logvaldiffs(k) -= a_(ls_ * sf + oldtilde);
-          logvaldiffs(k) += a_(ls_ * sf + newtilde);
-
-          thetasnew_ -= W_.row(ls_ * sf + oldtilde);
-          thetasnew_ += W_.row(ls_ * sf + newtilde);
-        }
-
-        RbmSpin::lncosh(thetasnew_, lnthetasnew_);
-        logvaldiffs(k) += lnthetasnew_.sum() - logtsum;
-      }
-    }
-    return logvaldiffs;
-  }
-
+      const std::vector<std::vector<double>> &newconf) override;
   // Difference between logarithms of values, when one or more visible variables
   // are being changed Version using pre-computed look-up tables for efficiency
   // on a small number of local changes
-  Complex LogValDiff(VisibleConstType v, const std::vector<int> &tochange,
-                     const std::vector<double> &newconf,
-                     const LookupType &lt) override {
-    Complex logvaldiff = 0.;
+  virtual Complex LogValDiff(VisibleConstType v,
+                             const std::vector<int> &tochange,
+                             const std::vector<double> &newconf,
+                             const LookupType &lt) override;
 
-    if (tochange.size() != 0) {
-      RbmSpin::lncosh(lt.V(0), lnthetas_);
-
-      thetasnew_ = lt.V(0);
-
-      for (std::size_t s = 0; s < tochange.size(); s++) {
-        const int sf = tochange[s];
-        const int oldtilde = confindex_[v[sf]];
-        const int newtilde = confindex_[newconf[s]];
-
-        logvaldiff -= a_(ls_ * sf + oldtilde);
-        logvaldiff += a_(ls_ * sf + newtilde);
-
-        thetasnew_ -= W_.row(ls_ * sf + oldtilde);
-        thetasnew_ += W_.row(ls_ * sf + newtilde);
-      }
-
-      RbmSpin::lncosh(thetasnew_, lnthetasnew_);
-      logvaldiff += (lnthetasnew_.sum() - lnthetas_.sum());
-    }
-    return logvaldiff;
-  }
+ private:
+  inline void Init();
 
   // Computhes the values of the theta pseudo-angles
   inline void ComputeTheta(VisibleConstType v, VectorType &theta) {
@@ -356,72 +123,8 @@ class RbmMultival : public AbstractMachine {
     vtilde = t.template cast<double>();
   }
 
-  const AbstractHilbert &GetHilbert() const noexcept override {
-    return hilbert_;
-  }
-
-  void to_json(json &j) const override {
-    j["Name"] = "RbmMultival";
-    j["Nvisible"] = nv_;
-    j["Nhidden"] = nh_;
-    j["LocalSize"] = ls_;
-    j["UseVisibleBias"] = usea_;
-    j["UseHiddenBias"] = useb_;
-    j["a"] = a_;
-    j["b"] = b_;
-    j["W"] = W_;
-  }
-
-  void from_json(const json &pars) override {
-    if (pars.at("Name") != "RbmMultival") {
-      throw InvalidInputError(
-          "Error while constructing RbmMultival from Json input");
-    }
-
-    if (FieldExists(pars, "Nvisible")) {
-      nv_ = pars["Nvisible"];
-    }
-
-    if (nv_ != hilbert_.Size()) {
-      throw InvalidInputError(
-          "Loaded wave-function has incompatible Hilbert space");
-    }
-
-    if (FieldExists(pars, "LocalSize")) {
-      ls_ = pars["LocalSize"];
-    }
-    if (ls_ != hilbert_.LocalSize()) {
-      throw InvalidInputError(
-          "Loaded wave-function has incompatible Hilbert space");
-    }
-
-    if (FieldExists(pars, "Nhidden")) {
-      nh_ = FieldVal(pars, "Nhidden");
-    } else {
-      nh_ = nv_ * double(FieldVal(pars, "Alpha"));
-    }
-
-    usea_ = FieldOrDefaultVal(pars, "UseVisibleBias", true);
-    useb_ = FieldOrDefaultVal(pars, "UseHiddenBias", true);
-
-    Init();
-
-    // Loading parameters, if defined in the input
-    if (FieldExists(pars, "a")) {
-      a_ = pars["a"];
-    } else {
-      a_.setZero();
-    }
-
-    if (FieldExists(pars, "b")) {
-      b_ = pars["b"];
-    } else {
-      b_.setZero();
-    }
-    if (FieldExists(pars, "W")) {
-      W_ = pars["W"];
-    }
-  }
+  virtual void to_json(json &j) const override;
+  virtual void from_json(const json &pars) override;
 };
 
 }  // namespace netket
